reorganization in process, not functional currently
