+++ Cogs/Queue.py
@@ -1,0 +1,6 @@
+from nextcord.ext import commands
+
+
+class Queue(commands.Cog):
+    def __init__(self, bot):
+        self.bot = bot
